moved validation to its own source file
